Improve ZipUtils performance and performance of loadClass
Before we were waiting for exceptions to be thrown so that we could deem the file as a zip.
However that solution is very bad so instead we check if the first 4 bytes of the file matches the zip magic.
If it does we plop it into our cache and continue with our day!

We also do the same for loadClass except ask the classloader to provide us the data. Gotta rely on others sometimes